requirements.txt
google-adk
uvicorn[standard]
fastapi